Fix: Data typee on preference request

+++ pkg/property-service/property-handlers/property-handlers.go
@@ -18,7 +18,7 @@ type PrefRequest struct {
 	LOCATIONS     []string `json:"locations"`
 	BUDGET        string   `json:"budget"`
 	BEDROOMS      uint     `json:"bedrooms"`
-	PROPERTY_SIZE uint     `json:"property_size"`
+	PROPERTY_SIZE float64  `json:"property_size"`
 	AMENITIES     []string `json:"amenities"`
 }
